backend-jumbo-jcldev/OrderService/src/main/java/com/cencosud/middleware/order/model/JumboItemDetail.java
package com.cencosud.middleware.order.model;

public class JumboItemDetail extends ItemDetail{
	
	private String productReference;
	private String brand;
	private String measurementUnit;
	private String skuId;
	
	public String getProductReference() {
		return productReference;
	}

	public void setProductReference(String productReference) {
		this.productReference = productReference;
	}

	public String getBrand() {
		return brand;
	}

	public void setBrand(String brand) {
		this.brand = brand;
	}

	public String getMeasurementUnit() {
		return measurementUnit;
	}

	public void setMeasurementUnit(String measurementUnit) {
		this.measurementUnit = measurementUnit;
	}

	public String getSkuId() {
		return skuId;
	}

	public void setSkuId(String skuId) {
		this.skuId = skuId;
	}
}
